UI update and paging added on HomeScreen

+++ app/src/main/java/bek/droid/news/data/model/response/BaseResponse.kt
@@ -2,6 +2,6 @@ package bek.droid.news.data.model.response
 
 class BaseResponse<T>(
     val status: String,
-    val totalResults: Long,
+    val totalResults: Int,
     val articles: List<T>
 )
